predictionServer/requirements.txt
django>=3.2.0
djangorestframework>=3.12.0
drf-yasg>=1.21.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
joblib>=1.0.0
psycopg2-binary>=2.9.0